Rename publish_course_manual_rb to course_manual_publisher, replacing the bash version; support rudimentary relative path normalization
