+++ examples/course_manual_publisher/ABC_training_course_manual.html
