+++ examples/course_manual_publisher/Dockerfile
