+++ examples/course_manual_publisher/Gemfile
@@ -8,6 +8,7 @@ git_source(:github) {|repo_name| "https://github.com/#{repo_name}" }
 gem "hashie", "~> 5.0.0"
 gem "httplog", "~> 1.6.2"
 gem "mime-types", "~> 3.4.1"
+gem "nokogiri", "~> 1.14.2"
 gem "rest-client", "~> 2.0.2"
-gem "slop", "~> 4.9.3"
 gem "activesupport", "~> 7.0.4"
+gem "pry"
