+++ examples/course_manual_publisher/README.md
