+++ examples/course_manual_publisher/files/ABC Settings.csv
