+++ examples/course_manual_publisher/files/ABC Training Manual.pdf
