+++ examples/course_manual_publisher/files/MainMenu.png
