+++ examples/course_manual_publisher/files/Notes on Product ABC.docx
