--- examples/publish_course_manual/README.md
@@ -1,27 +1,0 @@
-# Course Manager Publish Course Manual script
-
-## How this script works
-
-This script uses the Skytap Course Manager API to upload a manual and its attachments.
-
-To authenticate to Course Manager, the script needs an API key and secret generated in Course Manager.
-
-## Getting started
-
-Edit the variables at the top of the publish_manual.sh script and run it locally or as part of an automated action.
-
-## License
-
-Copyright 2022 Skytap Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-<http://www.apache.org/licenses/LICENSE-2.0>
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
