--- examples/publish_course_manual_rb/files/ABC Settings.csv
@@ -1,19 +1,0 @@
-type,id,course_name,region_name,user_identifier,user_url,skytap_configuration_skytap_id,created_at_utc,skytap_configuration_created_at_utc,discarded_at_utc,tag_names,event_id,event_name,user_first_name,user_last_name,user_passcode,event_created_at_utc,event_start_at_utc,event_end_at_utc,explicit_end_at_utc,subscriber_type,subscriber_email,integration,consumed_at_utc,skytap_object_skytap_id,integration_customer_name,integration_lab_purpose,integration_auth_token,integration_purpose,integration_description,integration_demo_type,integration_opp_id,integration_internal_sfdc_account_id,integration_internal_sfdc_opp_id,integration_internal_sfdc_requestors_user_id,integration_company_name,integration_provision_strategy,integration_post_provision_webhook_url,integration_post_creation_notification_email,integration_message_id,integration_webhook_token,integration_subject,integration_to,integration_cc,integration_bcc,integration_from,integration_text,integration_spf,snapshot_name
-Instructor,322,CF-110 XYZ Fundamentals,US-West,ssavoor+i@skytap.com,https://cloud.skytap.com/vms/a285c51f8fddf3ce565bf2c590ad71c8/desktops,92391494,2020-08-01 13:15:12,2020-08-01 13:15:13,,,102,Pre release 1019,,,4QYDV44KBU3G,2020-08-01 13:14:31,2020-08-03 13:00:00,2020-08-03 18:00:00,,,,,,92391494,,,,,,,,,,,,,,,,,,,,,,,,
-Instructor,325,1017 ILT,US-East,ssavoor+i@skytap.com,https://cloud.skytap.com/vms/57581a29fea4aa80ce0fa4973454185f/desktops,92398398,2020-08-01 15:18:28,2020-08-01 15:18:28,,,103,New ASAP 8/01,,,HCQ9Z4PW3RTR,2020-08-01 15:18:08,,2020-08-01 17:30:00,,,,,,92398398,,,,,,,,,,,,,,,,,,,,,,,,
-Lab,195,Practice Lab for CC-23 (Advanced),US-West,ssavoor@skytap.com,https://cloud.skytap.com/vms/06f69f377820b25daeef29a1e148492f/desktops,92395586,2020-08-01 13:09:20,2020-08-01 13:09:20,,,,,,,,,,,,,skytap.com,email,,92395586,,,,"",,,,,,,,,,,,,,,,,,,,
-Lab,196,Practice Lab for CC-23 (Advanced),EMEA,ss moodle,https://cloud.skytap.com/vms/1fb8f7413e91d7f31b927306828e9d6e/desktops,92395670,2020-08-01 13:09:47,2020-08-01 13:09:47,,Expiry 1 hour,,,,,,,,,,,,moodle,2020-08-01 16:00:59,92395670,,,"",,,,,,,,,,,,,,,,,,,,,
-Lab,197,Is an SL 1018,US-East,ss pre release 2 hr exp,https://cloud.skytap.com/vms/fd7bebe2066f9a4297829a447e6f59e2/desktops,92391202,2020-08-01 13:11:20,2020-08-01 13:11:20,,Expiry 2 hours,,,,,,,,,,,,basic,2020-08-01 16:01:02,92391202,"","",,,,,,,,,,,,,,,,,,,,,,
-Lab,198,Is an SL 1018,US-East,ss pre release 3 hr exp,https://cloud.skytap.com/vms/1028a9309f37ec0209ddf41d4d03dde8/desktops,92391222,2020-08-01 13:11:45,2020-08-01 13:11:45,,Expiry 3 hours,,,,,,,,,,,,basic,,92391222,"","",,,,,,,,,,,,,,,,,,,,,,
-Lab,199,Is an SL 1018,US-East,ss pre release 2 day exp,https://cloud.skytap.com/vms/7b4f2120852f94e37ea053bc8b01cfb0/desktops,92391226,2020-08-01 13:12:03,2020-08-01 13:12:03,,"Idle Expiry 1 day,Expiry 2 days",,,,,,,,,,,,basic,,92391226,"","",,,,,,,,,,,,,,,,,,,,,,
-Lab,200,Pre 1019 with credentials,US-West,with credset1,https://cloud.skytap.com/vms/00825d963ec35b34dac5ffbdc58d8318/desktops,92391670,2020-08-01 13:19:47,2020-08-01 13:19:47,,,,,,,,,,,,,,basic,,92391670,"","",,,,,,,,,,,,,,,,,,,,,,
-Lab,201,Pre 1019 with credentials,US-West,ssavoor@skytap.com,https://cloud.skytap.com/vms/9235bc096413e1ff3865992c76bdee36/desktops,92395574,2020-08-01 13:21:21,2020-08-01 13:21:21,,,,,,,,,,,,,skytap.com,email,,92395574,,,,"",,,,,,,,,,,,,,,,,,,,
-Lab,202,Another pre 1019,US-East,ssavoor@skytap.com,https://cloud.skytap.com/vms/75083027a195377c94d4d1b02161d50d/desktops,92395578,2020-08-01 13:22:26,2020-08-01 13:22:26,,Expiry 1 hour,,,,,,,,,,,skytap.com,email,2020-08-01 16:01:14,92395578,,,,"",,,,,,,,,,,,,,,,,,,,
-Lab,203,New creds course,US-East-2,ssavoor@skytap.com,https://cloud.skytap.com/vms/c83bab4f04d21f184a73c836b8724334/desktops,92396650,2020-08-01 14:49:54,2020-08-01 14:49:54,,Expiry 1 hour,,,,,,,,,,,skytap.com,email,2020-08-01 16:01:18,92396650,,,,"",,,,,,,,,,,,,,,,,,,,
-Lab,204,New creds course,US-East-2,grbiufns,https://cloud.skytap.com/vms/aea5516bf695ff10bc0e5fb8a610a0e8/desktops,92396674,2020-08-01 14:50:37,2020-08-01 14:50:38,,,,,,,,,,,,,,moodle,,92396674,,,"",,,,,,,,,,,,,,,,,,,,,
-Lab,205,Is an SL 1018,US-East,SS api course tag not applied,https://cloud.skytap.com/vms/04d965c57dc0edfba80778415df8fe2f/desktops,92398198,2020-08-01 15:14:35,2020-08-01 15:14:35,,Prov Expiry 1 day,,,,,,,,,,,,api,,92398198,,,,,,,,,,,,Delayed,https://example.com/posthook,ssavoor@skytap.com,,,,,,,,,,
-Student,323,CF-110 XYZ Fundamentals,US-West,ssavoor+s@skytap.com,https://cloud.skytap.com/vms/428acb9b23058f5ebd48b0c21785b063/desktops,92391486,2020-08-01 13:15:12,2020-08-01 13:15:12,,,102,Pre release 1019,,,BJFKRMGM5J5F,2020-08-01 13:14:31,2020-08-03 13:00:00,2020-08-03 18:00:00,,,,,,92391486,,,,,,,,,,,,,,,,,,,,,,,,
-Student,324,CF-110 XYZ Fundamentals,US-West,Student324,https://cloud.skytap.com/vms/ed3164db4f145cc90829439bdde82b89/desktops,92391406,2020-08-01 13:15:12,2020-08-01 13:15:13,,,102,Pre release 1019,,,HJFQNDC4CSEZ,2020-08-01 13:14:31,2020-08-03 13:00:00,2020-08-03 18:00:00,,,,,,92391406,,,,,,,,,,,,,,,,,,,,,,,,
-Student,326,1017 ILT,US-East,ssavoor+s@skytap.com,https://cloud.skytap.com/vms/0e958e198b3b9b0c56c91db21e527561/desktops,92398318,2020-08-01 15:18:28,2020-08-01 15:18:28,,,103,New ASAP 8/01,Shweta,"",XCT4S7XPN8VR,2020-08-01 15:18:08,,2020-08-01 17:30:00,,,,,,92398318,,,,,,,,,,,,,,,,,,,,,,,,
-Student,327,1017 ILT,US-East,Student327,https://cloud.skytap.com/vms/bb01127d7a5594e947fd2691f221468c/desktops,92398322,2020-08-01 15:18:28,2020-08-01 15:18:28,,,103,New ASAP 8/01,,,79U3TGBSYEJY,2020-08-01 15:18:08,,2020-08-01 17:30:00,,,,,,92398322,,,,,,,,,,,,,,,,,,,,,,,,
-Student,328,1017 ILT,US-West,Student328,https://cloud.skytap.com/vms/c3fd69da3a50aaa6f5bfe62a25e612a5/desktops,92401286,2020-08-01 16:10:59,2020-08-01 16:10:59,,,103,New ASAP 8/01,,,5W2YNMDJU4HW,2020-08-01 15:18:08,,2020-08-01 17:30:00,,,,,,92401286,,,,,,,,,,,,,,,,,,,,,,,,
